requirements.txt
streamlit
finance-datareader
pykrx
pandas
numpy
plotly
supabase
python-dotenv
